# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AMUPF Digital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List incident reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: department_id
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit an incident report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    patch:
      operationId: updateReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/status:
    patch:
      operationId: updateReportStatus
      tags: [reports]
      summary: Update report status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/feedback:
    post:
      operationId: submitReportFeedback
      tags: [reports]
      summary: Submit student feedback on a resolved report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "200":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # --- MEMBERS ---
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: assigned_to
          in: query
          schema:
            type: string
        - name: department_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Assign a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- ANNOUNCEMENTS ---
...
components:
  schemas:
...
    Report:
      type: object
      required: [report_id, department_id, title, description, location, incident_date, category, status, is_anonymous, created_at]
      properties:
        report_id:
          type: string
        student_id:
          type: ["string", "null"]
        student_name:
          type: ["string", "null"]
        department_id:
          type: integer
        department_name:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        location:
          type: string
        incident_date:
          type: string
        category:
          type: string
          enum: [physical, verbal, property, mental_health, other]
        status:
          type: string
          enum: [pending, in_progress, on_hold, resolved, rejected, escalated]
        evidence_urls:
          type: array
          items:
            type: string
        is_anonymous:
          type: boolean
        resolved_by:
          type: ["string", "null"]
        resolution_summary:
          type: ["string", "null"]
        rejection_reason:
          type: ["string", "null"]
        executive_notes:
          type: ["string", "null"]
        president_notes:
          type: ["string", "null"]
        feedback_rating:
          type: ["integer", "null"]
        feedback_comment:
          type: ["string", "null"]
        created_at:
          type: string
    ReportInput:
      type: object
      required: [department_id, title, description, location, incident_date, category]
      properties:
        department_id:
          type: integer
        title:
          type: string
          minLength: 10
          maxLength: 200
        description:
          type: string
          minLength: 50
          maxLength: 2000
        location:
          type: string
        incident_date:
          type: string
        category:
          type: string
          enum: [physical, verbal, property, mental_health, other]
        is_anonymous:
          type: boolean
        evidence_urls:
          type: array
          items:
            type: string
    ReportUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        executive_notes:
          type: string
        president_notes:
          type: string
        resolution_summary:
          type: string
    ReportStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, in_progress, on_hold, resolved, rejected, escalated]
        reason:
          type: string
        resolution_summary:
          type: string
        executive_notes:
          type: string
    FeedbackInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # --- MEMBER ---
...
    Task:
      type: object
      required: [task_id, title, description, assigned_to, assigned_by, priority, status, created_at]
      properties:
        task_id:
          type: string
        title:
          type: string
        description:
          type: string
        assigned_to:
          type: string
        assigned_to_name:
          type: ["string", "null"]
        assigned_by:
          type: string
        assigned_by_name:
          type: ["string", "null"]
        department_id:
          type: ["integer", "null"]
        department_name:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        due_date:
          type: ["string", "null"]
        related_report_id:
          type: ["string", "null"]
        completed_at:
          type: ["string", "null"]
        created_at:
          type: string
    TaskInput:
      type: object
      required: [title, description, assigned_to, priority]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        assigned_to:
          type: string
        department_id:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        due_date:
          type: string
        related_report_id:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        priority:
          type: string
        due_date:
          type: string

    # --- ANNOUNCEMENT ---
...